+++ src/fums/shaders.jl
@@ -4,74 +4,84 @@ module Shaders
 
 import Fae.@fum
 
-previous = @fum function previous()
+previous = @fum color function previous()
+    return (red, green, blue, alpha)
 end
 
 # for now, to force a color, just make each color channel = -channel + 2*color
 # used only for final colors
-force_red = @fum function force_red()
+force_red = @fum color function force_red()
     red = -red + 2
     green = -green
     blue = -blue
     alpha = -alpha + 2
+    return (red, green, blue, alpha)
 end
 
-custom = @fum function custom(; red = 0, green = 0, blue = 0, alpha = 0)
+custom = @fum color function custom(; r = 0, g = 0, b = 0, a = 0)
     red = red
     green = green
     blue = blue
     alpha = alpha
+    return (red, green, blue, alpha)
 end
 
-gray = @fum function gray()
+gray = @fum color function gray()
     red = 0.5
     green = 0.5
     blue = 0.5
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
 grey = gray
 
-red = @fum function red()
+red = @fum color function red()
     red = 1
     green = 0
     blue = 0
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
-green = @fum function green()
+green = @fum color function green()
     red = 0
     green = 1
     blue = 0
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
-blue = @fum function blue()
+blue = @fum color function blue()
     red = 0
     green = 0
     blue = 1
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
-magenta = @fum function magenta()
+magenta = @fum color function magenta()
     red = 1
     green = 0
     blue = 1
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
-white = @fum function white()
+white = @fum color function white()
     red = 1
     green = 1
     blue = 1
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
-black = @fum function black()
+black = @fum color function black()
     red = 0
     green = 0
     blue = 0
     alpha = 1
+    return (red, green, blue, alpha)
 end
 
 end
